[opt] ImageViewer component refactoring #199
