--- src/components/annotator/CategoriesList/ImageViewerList/ImageViewerList.tsx
@@ -1,333 +1,0 @@
-import Drawer from "@mui/material/Drawer";
-import ListItem from "@mui/material/ListItem";
-import ListItemText from "@mui/material/ListItemText";
-import React from "react";
-import { useNavigate } from "react-router-dom";
-import { Category, CategoryType } from "types/Category";
-import {
-  imageSelector,
-  unknownAnnotationCategorySelector,
-} from "store/selectors";
-import { batch, useDispatch, useSelector } from "react-redux";
-import { CollapsibleList } from "../CollapsibleList";
-import ListItemSecondaryAction from "@mui/material/ListItemSecondaryAction";
-import IconButton from "@mui/material/IconButton";
-import MoreHorizIcon from "@mui/icons-material/MoreHoriz";
-import { useDialog } from "hooks";
-import { useTranslation } from "hooks/useTranslation";
-import { Chip, Divider, Tooltip } from "@mui/material";
-import List from "@mui/material/List";
-import DeleteIcon from "@mui/icons-material/Delete";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import FolderOpenIcon from "@mui/icons-material/FolderOpen";
-import SaveIcon from "@mui/icons-material/Save";
-import Typography from "@mui/material/Typography";
-import Toolbar from "@mui/material/Toolbar";
-import AppBar from "@mui/material/AppBar";
-import PopupState, { bindTrigger } from "material-ui-popup-state";
-import { selectedAnnotationsIdsSelector } from "store/selectors/selectedAnnotationsIdsSelector";
-import ListItemAvatar from "@mui/material/ListItemAvatar";
-import Avatar from "@mui/material/Avatar";
-import { ImageMenu } from "../ImageMenu";
-import { DeleteAllAnnotationsDialog } from "../DeleteAllAnnotationsDialog";
-import { SaveMenu } from "../SaveMenu/SaveMenu";
-import { OpenMenu } from "../OpenMenu/OpenMenu";
-import { AnnotatorHelpDrawer } from "components/common/Help";
-import { imageViewerSlice, setActiveImage } from "store/slices";
-import { ImageType, ShadowImageType } from "types/ImageType";
-import { ArrowBack } from "@mui/icons-material";
-import { annotatorImagesSelector } from "store/selectors/annotatorImagesSelector";
-import { createdAnnotatorCategoriesSelector } from "store/selectors/createdAnnotatorCategoriesSelector";
-import { ExitAnnotatorDialog } from "../ExitAnnotatorDialog";
-import { AppBarOffset } from "components/styled/AppBarOffset";
-import { SendFeedbackListItem } from "components/common/SendFeedbackListItem";
-import { LogoIcon } from "components/Logo";
-import { CategoriesList } from "components/CategoriesList";
-
-export const ImageViewerList = () => {
-  const createdCategories = useSelector(createdAnnotatorCategoriesSelector);
-  const unknownAnnotationCategory = useSelector(
-    unknownAnnotationCategorySelector
-  );
-
-  const selectedAnnotationsIds = useSelector(selectedAnnotationsIdsSelector);
-
-  const currentImage = useSelector(imageSelector);
-
-  const annotatorImages = useSelector(annotatorImagesSelector);
-
-  const dispatch = useDispatch();
-
-  const navigate = useNavigate();
-
-  const {
-    onClose: onCloseDeleteAllAnnotationsDialog,
-    onOpen: onOpenDeleteAllAnnotationsDialog,
-    open: openDeleteAllAnnotationsDialog,
-  } = useDialog();
-
-  const {
-    onClose: onCloseExitAnnotatorDialog,
-    onOpen: onOpenExitAnnotatorDialog,
-    open: openExitAnnotatorDialog,
-  } = useDialog();
-
-  const [imageAnchorEl, setImageAnchorEl] = React.useState<null | HTMLElement>(
-    null
-  );
-
-  const [selectedImage, setSelectedImage] = React.useState<ImageType>(
-    currentImage!
-  );
-
-  const onReturnToMainProject = () => {
-    onCloseExitAnnotatorDialog();
-    navigate("/");
-  };
-
-  const onCategoryClickCallBack = (category: Category) => {
-    dispatch(
-      imageViewerSlice.actions.setSelectedCategoryId({
-        selectedCategoryId: category.id,
-      })
-    );
-  };
-
-  const onImageMenuOpen = (
-    event: React.MouseEvent<HTMLButtonElement>,
-    image: ShadowImageType
-  ) => {
-    setImageAnchorEl(event.currentTarget);
-    setSelectedImage(image as ImageType);
-  };
-
-  const onImageMenuClose = () => {
-    setImageAnchorEl(null);
-  };
-
-  const onClearAllAnnotations = () => {
-    const existingAnnotations = annotatorImages
-      .map((image: ShadowImageType) => {
-        return [...image.annotations];
-      })
-      .flat();
-    if (existingAnnotations.length) {
-      onOpenDeleteAllAnnotationsDialog();
-    }
-  };
-
-  const onClearSelectedAnnotations = () => {
-    if (!selectedAnnotationsIds) return;
-    batch(() => {
-      dispatch(
-        imageViewerSlice.actions.deleteImageInstances({
-          ids: selectedAnnotationsIds,
-        })
-      );
-      dispatch(
-        imageViewerSlice.actions.setSelectedCategoryId({
-          selectedCategoryId: unknownAnnotationCategory.id,
-        })
-      );
-      dispatch(
-        imageViewerSlice.actions.setSelectedAnnotations({
-          selectedAnnotations: [],
-          selectedAnnotation: undefined,
-        })
-      );
-    });
-  };
-
-  const onImageItemClick = (
-    evt: React.MouseEvent<HTMLDivElement | HTMLButtonElement, MouseEvent>,
-    image: ShadowImageType
-  ) => {
-    dispatch(setActiveImage({ imageId: image.id }));
-  };
-
-  const t = useTranslation();
-
-  return (
-    <Drawer
-      anchor="left"
-      sx={{
-        flexShrink: 0,
-        width: (theme) => theme.spacing(32),
-        "& > .MuiDrawer-paper": {
-          width: (theme) => theme.spacing(32),
-        },
-      }}
-      open
-      variant="persistent"
-    >
-      <AppBar
-        color="default"
-        sx={{
-          backgroundColor: "rgba(0, 0, 0, 0)",
-          boxShadow: "none",
-          position: "absolute",
-        }}
-      >
-        <Toolbar>
-          <Tooltip title="Save and return to project" placement="bottom">
-            <IconButton
-              edge="start"
-              onClick={onOpenExitAnnotatorDialog}
-              aria-label="Exit Annotator"
-              href={""}
-            >
-              <ArrowBack />
-            </IconButton>
-          </Tooltip>
-          <LogoIcon width={30} height={30} />
-          <Typography variant="h5" color={"#02aec5"}>
-            Annotator
-          </Typography>
-        </Toolbar>
-      </AppBar>
-
-      <AppBarOffset />
-
-      <ExitAnnotatorDialog
-        onReturnToProject={onReturnToMainProject}
-        onClose={onCloseExitAnnotatorDialog}
-        open={openExitAnnotatorDialog}
-      />
-
-      <Divider />
-
-      <List dense>
-        <OpenListItem />
-        <SaveListItem />
-      </List>
-
-      <Divider />
-
-      <CollapsibleList closed dense primary={t("Images")}>
-        {annotatorImages.map((image: ShadowImageType) => {
-          return (
-            <div key={image.id}>
-              <ListItem
-                button
-                id={image.id}
-                onClick={(evt: React.MouseEvent<HTMLDivElement, MouseEvent>) =>
-                  onImageItemClick(evt, image)
-                }
-                selected={image.id === currentImage?.id}
-              >
-                <ListItemAvatar>
-                  <Avatar alt={image.name} src={image.src} variant={"square"} />
-                </ListItemAvatar>
-                <ListItemText
-                  id={image.id}
-                  primary={image.name}
-                  primaryTypographyProps={{ noWrap: true }}
-                />
-                {image.annotations.length !== 0 && (
-                  <Chip label={image.annotations.length} size="small" />
-                )}
-                <ListItemSecondaryAction>
-                  <IconButton
-                    edge="end"
-                    onClick={(event) => onImageMenuOpen(event, image)}
-                  >
-                    <MoreHorizIcon />
-                  </IconButton>
-                </ListItemSecondaryAction>
-              </ListItem>
-            </div>
-          );
-        })}
-        <ImageMenu
-          anchorElImageMenu={imageAnchorEl}
-          selectedImage={selectedImage}
-          onCloseImageMenu={onImageMenuClose}
-          openImageMenu={Boolean(imageAnchorEl)}
-        />
-      </CollapsibleList>
-
-      <Divider />
-
-      <CategoriesList
-        createdCategories={createdCategories}
-        unknownCategory={unknownAnnotationCategory}
-        predicted={false}
-        categoryType={CategoryType.AnnotationCategory}
-        onCategoryClickCallBack={onCategoryClickCallBack}
-      />
-
-      <Divider />
-
-      <List dense>
-        <ListItem button onClick={onClearAllAnnotations}>
-          <ListItemIcon>
-            <DeleteIcon color="disabled" />
-          </ListItemIcon>
-          <ListItemText primary={t("Clear all annotations")} />
-        </ListItem>
-
-        <DeleteAllAnnotationsDialog
-          onClose={onCloseDeleteAllAnnotationsDialog}
-          open={openDeleteAllAnnotationsDialog}
-        />
-
-        <ListItem button onClick={onClearSelectedAnnotations}>
-          <ListItemIcon>
-            <DeleteIcon color="disabled" />
-          </ListItemIcon>
-          <ListItemText primary={t("Clear selected annotations")} />
-        </ListItem>
-      </List>
-
-      <Divider />
-
-      <List dense>
-        <SendFeedbackListItem />
-        <AnnotatorHelpDrawer />
-      </List>
-    </Drawer>
-  );
-};
-
-const OpenListItem = () => {
-  return (
-    <PopupState variant="popover">
-      {(popupState) => (
-        <>
-          <ListItem button {...bindTrigger(popupState)}>
-            <ListItemIcon>
-              <FolderOpenIcon />
-            </ListItemIcon>
-
-            <ListItemText primary="Open" />
-          </ListItem>
-
-          <OpenMenu popupState={popupState} />
-        </>
-      )}
-    </PopupState>
-  );
-};
-
-const SaveListItem = () => {
-  const t = useTranslation();
-
-  return (
-    <>
-      <PopupState variant="popover">
-        {(popupState) => (
-          <>
-            <ListItem button {...bindTrigger(popupState)}>
-              <ListItemIcon>
-                <SaveIcon />
-              </ListItemIcon>
-
-              <ListItemText primary={t("Save")} />
-            </ListItem>
-            <SaveMenu popupState={popupState} />
-          </>
-        )}
-      </PopupState>
-    </>
-  );
-};
