--- src/components/annotator/CategoriesList/ImageViewerList/index.ts
@@ -1,1 +1,0 @@
-export { ImageViewerList } from "./ImageViewerList";
