--- src/components/annotator/CategoriesList/index.ts
@@ -1,1 +1,0 @@
-export { ImageViewerList } from "./ImageViewerList";
