+++ src/components/annotator/ImageViewerList/ClearAnnotations/index.ts
@@ -1,0 +1,1 @@
+export { ClearAnnotationsListItem } from "./ClearAnnotationsListItem";
