+++ src/components/annotator/ImageViewerList/ColorIcon/ColorIcon.tsx
