+++ src/components/annotator/ImageViewerList/ColorIcon/index.ts
