+++ src/components/annotator/ImageViewerList/DeleteAllAnnotationsDialog/DeleteAllAnnotationsDialog.tsx
