+++ src/components/annotator/ImageViewerList/DeleteAllAnnotationsDialog/index.ts
