+++ src/components/annotator/ImageViewerList/ExampleImageDialog/OpenExampleImageMenuItem.tsx
