+++ src/components/annotator/ImageViewerList/ImageList/index.ts
@@ -1,0 +1,1 @@
+export { ImageList } from "./ImageList";
