+++ src/components/annotator/ImageViewerList/ImageViewerAppBar/index.ts
@@ -1,0 +1,1 @@
+export { ImageViewerAppBar } from "./ImageViewerAppBar";
