+++ src/components/annotator/ImageViewerList/ImageViewerDrawer/index.ts
@@ -1,0 +1,1 @@
+export { ImageViewerDrawer } from "./ImageViewerDrawer";
