+++ src/components/annotator/ImageViewerList/OpenMenu/OpenExampleImageMenuItem.tsx
