+++ src/components/annotator/ImageViewerList/index.ts
@@ -1,0 +1,1 @@
+export { ImageViewerDrawer } from "./ImageViewerDrawer";
